feat: add EdgeExclusionLayer component for crop screen
